solve the logout problem

+++ src/components/page/ssoLogout.vue
@@ -1,6 +1,6 @@
 <template>
   <div>
-</div>
+  </div>
 </template>
 
 <script>
@@ -12,16 +12,23 @@ export default {
   },
   methods:{
   },
-  mounted:function () {
-    this.$token.deleteToken();
-    var authorUrl = this.$config.userAuthorizationUri;
-    authorUrl = authorUrl + ('?' + this.$querystring.stringify({
-      client_id:this.$config.clientId,
-      response_type:this.$config.response_type,
-      state:this.$config.code,
-      redirect_uri:this.$config.redirect_uri, 
-    }))
-    window.location.href = authorUrl; 
+  mounted:function () { 
+    if(this.$token.loadToken().access_token !== "null"){      
+      this.$token.deleteToken();
+      window.location.href = "http://120.78.186.60:8087/logout"; 
+
+    } 
+    else {
+      var authorUrl = this.$config.userAuthorizationUri;
+      authorUrl = authorUrl + ('?' + this.$querystring.stringify({
+        client_id:this.$config.clientId,
+        response_type:this.$config.response_type,
+        state:'123',
+        redirect_uri:this.$config.redirect_uri, 
+      }))
+      window.location.href = authorUrl
+    }
+    
   }
 }
 </script>
